+++ metrics/api/src/main/java/io/helidon/metrics/api/Sample.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2022 Oracle and/or its affiliates.
+ * Copyright (c) 2021, 2023 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -48,7 +48,7 @@ public interface Sample {
      * @param value value
      * @return a new labeled value if exemplar handling is supported
      */
-    static Labeled labeled(long value) {
+    static Labeled labeled(double value) {
         return ExemplarServiceManager.isActive()
                 ? new LabeledSample(value, ExemplarServiceManager.exemplarLabel(), System.currentTimeMillis())
                 : new LabeledSample(value, ExemplarServiceManager.INACTIVE_LABEL, 0);
@@ -100,7 +100,7 @@ public interface Sample {
          * The value.
          * @return value
          */
-        long value();
+        double value();
 
         /**
          * Value label.
